requirements.txt
connexion == 1.1.15
python_dateutil == 2.6.0
setuptools >= 21.0.0
Flask_Cors == 3.0.3
pymysql == 0.9.3
flask-mysql == 1.4.0